refactor(components): reorganize imports and create new AppLayout, Sidebar, Breadcrumbs, and UserMenu components
