+++ components/common/index.ts
@@ -1,8 +1,4 @@
-export * from './sidebar';
 export * from './footer';
-export * from './user-menu';
 export * from './auth-button';
-export * from './breadcrumbs';
-export * from './app-layout';
 export * from './pagination';
 export * from './search-and-filter';
